// src/models/DescTsueldo.ts
export interface DescTsueldo {
  IdDescTsueldo: number;
  nameDescTsueldo: string | null;
  codeDescTsueldo: string | null;
}

export interface DescTsueldoCreate {
  nameDescTsueldo: string | null;
  codeDescTsueldo: string | null;
}

export interface DescTsueldoUpdate {
  nameDescTsueldo: string | null;
  codeDescTsueldo: string | null;
}